Complete the transition to EXI floats (10 based exponent)

+++ trunk/src/grammar/src/eventsEXI.c
@@ -52,12 +52,12 @@ errorCode writeEventCode(EXIStream* strm, GrammarRule* currentRule, unsigned cha
 	unsigned char i = 0;
 	for(i = 0; i < codeLength - 1; i++)
 	{
-		tmp_err_code = encodeNBitUnsignedInteger(strm, currentRule->bits[i], currentRule->prodCounts[i]);
+		tmp_err_code = encodeNBitUnsignedInteger(strm, currentRule->bits[i], (unsigned int) currentRule->prodCounts[i]);
 		if(tmp_err_code != ERR_OK)
 			return tmp_err_code;
 	}
 
-	tmp_err_code = encodeNBitUnsignedInteger(strm, currentRule->bits[codeLength - 1], codeLastPart);
+	tmp_err_code = encodeNBitUnsignedInteger(strm, currentRule->bits[codeLength - 1], (unsigned int) codeLastPart);
 	if(tmp_err_code != ERR_OK)
 		return tmp_err_code;
